Add "show files" tool that shows file contents python style
The output can be formatted with an "what-you-see-is-what-you-get" approch.
Simply provide a python evaluable file that is filled with a dict
containing the same keys(the filenames)
and instead of the data the format strings.
The intention is to be able to quicky edit and oversee a set of files
that are of short lenght.