add grr git-rebase-todo filter tool
The tool is able to filter git-rebase-todo files
and insert a short part in square brackets that
contains the modified file names in
uniq shortened and human readable form.
The insertion is at the start of the subject
and can be colored by your editor.
The shortening works by using
a fixed lenght of the filepath from
the end.
Then, if the file is not already unique, the
starting characters are replaced one by one
with the full filepathes sha1 hashs starting characters
, until uniqness is reached.
For bigger repos the lenght of the created
file-id should be increased to
make more of the path human readable.
Configure your editor to use the program to filter
git-rebase-todo and pass the repo location.
modified vim syntax file in my repo at:
https://github.com/alexander-n8hgeg5e/.nvim/blob/master/syntax/gitrebase.vim