flush stdout to prevent loosing the order if stdout is a pipe
There are buffers for printed things and things from subprocesses.
These parts go from the buffers into the stdout pipe, but the order
can get messed up.
Subprocess output should be flushed by itself when the subprocess returns.
So need to flush the stdout buffer that is used by print.